+++ hbase1/src/main/java/site/ycsb/db/hbase1/HBaseClient1.java
@@ -198,7 +198,7 @@ public class HBaseClient1 extends site.ycsb.DB {
       }
       long en = System.nanoTime();
       final String type = clientSideBuffering ? "UPDATE" : "CLEANUP";
-      measurements.measure(type, (int) ((en - st) / 1000));
+      measurements.measure(type, (long) ((en - st) / 1000));
       int threadCount = THREAD_COUNT.decrementAndGet();
       if (threadCount <= 0) {
         // Means we are done so ok to shut down the Connection.
